move opengsn to dev dependencies
