+++ package.json
@@ -54,7 +54,6 @@
   "homepage": "https://github.com/nevermined-io/sdk-js#readme",
   "dependencies": {
     "@nevermined-io/subgraphs": "0.4.1",
-    "@opengsn/provider": "^2.2.6",
     "assert": "^2.0.0",
     "codegen-graph-ts": "npm:@nevermined-io/codegen-graph-ts@0.1.4",
     "crypto-browserify": "^3.12.0",
@@ -75,6 +74,7 @@
   },
   "devDependencies": {
     "@faker-js/faker": "^6.3.1",
+    "@opengsn/provider": "^2.2.6",
     "@release-it/bumper": "^1.4.1",
     "@types/chai": "^4.2.12",
     "@types/chai-as-promised": "^7.1.3",
